AStarGridSearch

Documentation for AStarGridSearch.

AStarGridSearch.AStarGridSearch

AStarGridSearch.PathCost

AStarGridSearch.PathCost

AStarGridSearch.PathCost

AStarGridSearch.finalOrientation

AStarGridSearch.getPath

AStarGridSearch.AStarGridSearch — Module

This module contains types and functions for searching for paths on a 2D grid using A*. Its main use is to get the path cost (distance), but it can return the full path as well.

Main public types and functions:

PathCost

finalOrientation

getPath

AStarGridSearch.PathCost — Type

Struct for PathCost function data. Previous computations are kept track of in its data. Can be used multiple times for the same start cell, saving computation.

The cells of the costMatrix contain the distance to them from the start cell. NaN is a placeholder meaning no path has been calculated to that cell yet. Inf means that cell is not reachable from the start cell.

Fields:

start::Any

costMatrix::Any

resolution::Any

diffs::Any

dist::Any

heuristic::Any

frontier::Any

AStarGridSearch.PathCost — Method

PathCost(
    start,
    occupancy,
    resolution;
    diagonals,
    dist,
    heuristic
) -> PathCost

The constructor initializes the path search algorithm, created for each new start cell.

Inputs:

start: the starting cell in the matrix

occupancy: a matrix indicating which cells are occupied

resolution: a vector of the width and height of each cell

diagonals: whether to include connectivity to diagonal cells or not (default true)

dist: function to compute distance from current to next cell (default weightedEuclidian)

heuristic: function to compute estimated distance from current to goal cell (default weightedEuclidian)

Note: this type and associated methods can in fact be used with any N-dimensional cost array, not just a matrix.

AStarGridSearch.PathCost — Method

Search for a path to a desired goal cell using the A* algorithm.

Returns the path cost, which will be Inf if it is unreachable.

If the cells of the path are desired, use the getPath function.

AStarGridSearch.finalOrientation — Method

finalOrientation(S::PathCost, goal) -> Any

Given a PathCost and a goal point, this function returns the angle of the direction from penultimate cell to goal cell, effectively the direction at the end of the path to the goal.

AStarGridSearch.getPath — Method

getPath(S::PathCost, goal) -> Vector

Given a PathCost and a goal point, this function returns the entire list of cells from the start to the goal. Only useful to be called after the PathCost object is called with the goal cell.